chore(build): move event name annotation to standalone subproject
