+++ src/main/kotlin/net/ccbluex/liquidbounce/event/Event.kt
@@ -21,7 +21,7 @@ package net.ccbluex.liquidbounce.event
 import it.unimi.dsi.fastutil.objects.Object2ReferenceRBTreeMap
 import it.unimi.dsi.fastutil.objects.Reference2ObjectOpenHashMap
 import net.ccbluex.liquidbounce.config.gson.stategies.ProtocolExclude
-import net.ccbluex.liquidbounce.utils.client.Nameable
+import net.ccbluex.liquidbounce.annotations.Nameable
 
 /**
  * A callable event
