+++ src/main/kotlin/net/ccbluex/liquidbounce/event/events/DrawEvents.kt
@@ -21,7 +21,7 @@
 package net.ccbluex.liquidbounce.event.events
 
 import net.ccbluex.liquidbounce.event.Event
-import net.ccbluex.liquidbounce.utils.client.Nameable
+import net.ccbluex.liquidbounce.annotations.Nameable
 import net.minecraft.client.gl.Framebuffer
 import net.minecraft.client.gui.DrawContext
 import net.minecraft.client.render.Camera
