+++ src/main/kotlin/net/ccbluex/liquidbounce/event/events/WindowEvents.kt
@@ -23,7 +23,7 @@ package net.ccbluex.liquidbounce.event.events
 import net.ccbluex.liquidbounce.event.CancellableEvent
 import net.ccbluex.liquidbounce.event.Event
 import net.ccbluex.liquidbounce.integration.interop.protocol.event.WebSocketEvent
-import net.ccbluex.liquidbounce.utils.client.Nameable
+import net.ccbluex.liquidbounce.annotations.Nameable
 import net.minecraft.client.gui.screen.Screen
 import net.minecraft.client.util.InputUtil
